Added problem statement for project euler solutions

+++ src/main/java/org/open/corejava/projecteuler/ProjectEulerP6S1.java
@@ -2,6 +2,16 @@ package org.open.corejava.projecteuler;
 
 /**
  * @author Mahadev Mane
+ * <p>
+ * The sum of the squares of the first ten natural numbers is, 385
+ * <p>
+ * The square of the sum of the first ten natural numbers is, 3025
+ * <p>
+ * However, the difference between the sum of the squares of the
+ * first ten natural numbers and the square of the sum is 3025 - 385 = 2640.
+ * <p>
+ * Find the difference between the sum of the squares of the
+ * first one hundred natural numbers and the square of the sum.
  */
 
 public class ProjectEulerP6S1 {
@@ -10,7 +20,7 @@ public class ProjectEulerP6S1 {
         long sumOfSqr = 0, sqrOfSum;
 
         for (int i = lLimit; i <= uLimit; i++) {
-            sumOfSqr += (i * i);
+            sumOfSqr += ((long) i * i);
         }
 
         sqrOfSum = (uLimit * (uLimit + 1)) / 2;
